fix layout for meeting notes
