--- client/src/app/components/meeting-notes/notes-list/notes-list.component.html
@@ -1,24 +1,0 @@
-<div class="flex justify-between items-center">
-  <h2 class="text-xl font-semibold">
-    Meeting Notes
-  </h2>
-  <button
-    mat-icon-button
-    (click)="openUploadDialog()"
-    aria-label="Upload notes"
-  >
-    <mat-icon>add</mat-icon>
-  </button>
-</div>
-<mat-action-list>
-  @for (note of notesMetadata(); let i = $index; track note.id) {
-    <mat-list-item
-      (click)="openFile(note)"
-      class="!rounded-full"
-    >
-      <span matListItemTitle>
-        {{ note.name }}
-      </span>
-    </mat-list-item>
-  }
-</mat-action-list>
