+++ client/src/app/views/project-detail/project-detail.view.html
@@ -6,18 +6,18 @@
       Settings
     </button>
   </div>
-  <div class="flex flex-row gap-4 justify-center h-full flex-1">
-    <div class="flex flex-1 max-w-1/2 h-full flex-col gap-4 self-start">
+  <div class="flex-1 min-h-0 flex flex-row gap-4">
+    <div class="flex-1 min-h-0 w-1/2 flex flex-col gap-4">
       <app-meeting-notes
-        class="flex flex-col flex-1 h-1/2 p-4 outline !outline-(--mat-sys-outline-variant) rounded !shadow-(--mat-sys-level2)"
+        class="block h-1/2 min-h-0 p-4 outline !outline-(--mat-sys-outline-variant) rounded !shadow-(--mat-sys-level2)"
         [prelookCount]="3"
       ></app-meeting-notes>
       <app-report-overview
-        class="flex flex-col flex-1 h-1/2 p-4 outline !outline-(--mat-sys-outline-variant) rounded !shadow-(--mat-sys-level2)"
+        class="block h-1/2 min-h-0 p-4 outline !outline-(--mat-sys-outline-variant) rounded !shadow-(--mat-sys-level2)"
       ></app-report-overview>
     </div>
     <app-faq-chat
-      class="w-full max-w-1/2 p-4 outline !outline-(--mat-sys-outline-variant) rounded !shadow-(--mat-sys-level2) h-full"
+      class="w-1/2 p-4 outline !outline-(--mat-sys-outline-variant) rounded !shadow-(--mat-sys-level2) h-full"
     ></app-faq-chat>
   </div>
 </div>
