+++ client/src/app/views/report-overview/report-list/report-list.view.html
@@ -1,24 +1,26 @@
-@if (isLoading()) {
-<div class="flex items-center justify-center h-full">
-  <mat-progress-spinner mode="indeterminate"></mat-progress-spinner>
-</div>
-} @else if (reports().length === 0) {
-<span>No reports available.</span>
-} @else {
-<mat-action-list>
-  @for (report of reports(); track report.id) {
-  <button
-    mat-list-item
-    class="!rounded-full"
-    [disabled]="report.generating"
-    (click)="onReportSelected(report.id)"
-  >
-    <span class="flex items-center gap-2">
-      {{report.name}} @if (report.generating){
-      <mat-progress-spinner mode="indeterminate"></mat-progress-spinner>
-      }
-    </span>
-  </button>
+<div class="h-full flex flex-col">
+  @if (isLoading()) {
+  <div class=" h-full flex items-center justify-center">
+    <mat-progress-spinner mode="indeterminate"></mat-progress-spinner>
+  </div>
+  } @else if (reports().length === 0) {
+  <span>No reports available.</span>
+  } @else {
+  <mat-action-list>
+    @for (report of reports(); track report.id) {
+    <button
+      mat-list-item
+      class="!rounded-full"
+      [disabled]="report.generating"
+      (click)="onReportSelected(report.id)"
+    >
+      <span class="flex items-center gap-2">
+        {{report.name}} @if (report.generating){
+        <mat-progress-spinner mode="indeterminate"></mat-progress-spinner>
+        }
+      </span>
+    </button>
+    }
+  </mat-action-list>
   }
-</mat-action-list>
-}
+</div>
