+++ client/src/app/views/report-overview/report-overview.view.html
@@ -1,53 +1,51 @@
-<div class="flex gap-2 items-center">
-  <button
-    mat-icon-button
-    aria-label="Back to report overview"
-    [disabled]="!selectedReportId()"
-    (click)="selectedReportId.set(undefined)"
-  >
-    <mat-icon>arrow_back</mat-icon>
-  </button>
-  <h2 class="text-xl font-semibold">
-    {{selectedReport() ? selectedReport()!.name : "Reports"}}
-  </h2>
-</div>
-
-<!-- switch between list and content -->
-<div class="flex-1 overflow-y-auto">
-  @if (selectedReportId()){
-  <app-report-content [report]="selectedReport()!"></app-report-content>
-  } @else {
-  <app-report-list
-    (selectedReportId)="selectedReportId.set($event)"
-  ></app-report-list>
-  }
-</div>
-
-<!-- Always show the form below -->
-<div>
-  @if (!showForm()) {
-  <button
-    mat-mini-fab
-    color="primary"
-    aria-label="Add report"
-    (click)="displayForm()"
-  >
-    <mat-icon>add</mat-icon>
-  </button>
-  } @else {
-  <form
-    class="flex items-start justify-between rounded outline !outline-(--mat-sys-outline-variant) !shadow-(--mat-sys-level2) pt-4"
-    (ngSubmit)="onSubmit()"
-  >
+<div class="flex-1 flex flex-col h-full gap-4">
+  <mat-toolbar color="primary" class="flex justify-between">
+    <div class="flex flex-row gap-2 items-center">
+      <button
+        mat-icon-button
+        aria-label="Back to report overview"
+        [disabled]="!selectedReportId()"
+        (click)="selectedReportId.set(undefined)"
+      >
+        <mat-icon>arrow_back</mat-icon>
+      </button>
+      <span>
+        {{selectedReport() ? selectedReport()!.name : "Reports"}}
+      </span>
+    </div>
     <button
       mat-icon-button
-      type="button"
-      (click)="onCancel()"
-      aria-label="Cancel"
+      aria-label="Add report"
+      [matMenuTriggerFor]="addReportMenu"
+      (menuOpened)="openMenu()"
+      [matMenuTriggerRestoreFocus]="false"
     >
-      <mat-icon>close</mat-icon>
+      <mat-icon>add</mat-icon>
     </button>
-    <div class="flex flex-wrap gap-x-4">
+  </mat-toolbar>
+
+  <!-- switch between list and content -->
+  <div class="overflow-y-auto">
+    @if (selectedReportId()){
+    <app-report-content [report]="selectedReport()!"></app-report-content>
+    } @else {
+    <app-report-list
+      (selectedReportId)="selectedReportId.set($event)"
+    ></app-report-list>
+    }
+  </div>
+
+  <mat-menu
+    #addReportMenu="matMenu"
+    class="p-4"
+    xPosition="before"
+    (closed)="onCancel()"
+  >
+    <form
+      matMenuContent
+      class="flex flex-col gap-4"
+      (ngSubmit)="onSubmit(); addReportMenu.closed.emit()"
+    >
       <mat-form-field appearance="outline">
         <mat-label>Start Date</mat-label>
         <input
@@ -79,7 +77,7 @@
         ></mat-datepicker-toggle>
         <mat-datepicker #endPicker></mat-datepicker>
       </mat-form-field>
-      <mat-form-field appearance="outline">
+      <mat-form-field appearance="outline" (click)="$event.stopPropagation()">
         <mat-label>Users</mat-label>
         <mat-select multiple [(ngModel)]="userIds" name="userIds">
           @for (user of possibleUsers(); track user.id) {
@@ -87,10 +85,23 @@
           }
         </mat-select>
       </mat-form-field>
-    </div>
-    <button mat-icon-button color="primary" type="submit" aria-label="Send">
-      <mat-icon>send</mat-icon>
-    </button>
-  </form>
-  }
+      <div class="flex justify-end gap-2">
+        <button
+          matButton="outlined"
+          type="button"
+          (click)="onCancel(); addReportMenu.closed.emit()"
+        >
+          Cancel
+        </button>
+        <button
+          matButton="filled"
+          color="primary"
+          type="submit"
+          aria-label="Generate"
+        >
+          Generate
+        </button>
+      </div>
+    </form>
+  </mat-menu>
 </div>
