+++ client/src/app/views/sidebar/sidebar.component.html
@@ -53,7 +53,7 @@
       <mat-icon>chevron_right</mat-icon>
     </button>
     }
-    <div class="container p-4 mx-auto h-full">
+    <div class="p-4 mx-auto h-full">
       <router-outlet></router-outlet>
     </div>
   </mat-sidenav-content>
